Text: group and split test into logically object related files
- Create kitchen/tests/test_public_access_view.py, put tests relating accessing all pages with no logged in user

- Create kitchen/tests/test_custom_logout.py, put tests relating logging out user

- Create kitchen/tests/test_index_private_view.py, put tests relating private access index view

- Create kitchen/tests/test_cook_private_view.py, put tests relating private access cook views
